+++ structs/v2/shapes.go
@@ -1,9 +1,9 @@
 package main
 
-func Perimeter(width int, height int) (perimeter int) {
+func Perimeter(width float64, height float64) (perimeter float64) {
 	return 2 * (width + height)
 }
 
-func Area(width int, height int) (area int) {
+func Area(width float64, height float64) (area float64) {
 	return width * height
 }
