+++ structs/v3/shapes.go
@@ -1,14 +1,14 @@
 package main
 
 type Rectangle struct {
-	width  int
-	height int
+	width  float64
+	height float64
 }
 
-func Perimeter(rectangle Rectangle) (perimeter int) {
+func Perimeter(rectangle Rectangle) (perimeter float64) {
 	return 2 * (rectangle.width + rectangle.height)
 }
 
-func Area(rectangle Rectangle) (area int) {
+func Area(rectangle Rectangle) (area float64) {
 	return rectangle.width * rectangle.height
 }
